+++ CRUD en c#/MiPrimerCRUD/ManejoDeClientes.xaml.cs
@@ -106,24 +106,28 @@ namespace MiPrimerCRUD
 
         private void BtnBorrarCliente_Click(object sender, RoutedEventArgs e)
         {
-            try
+            MessageBoxResult messageBoxResult = System.Windows.MessageBox.Show("¿Quieres borrar al cliente seleccionado?", "Mensaje", System.Windows.MessageBoxButton.YesNo);
+            if (messageBoxResult == System.Windows.MessageBoxResult.Yes)
             {
-                // creamos una consulta parametrica para hacer posible la eliminacion de un registro
-                string consulta = "DELETE FROM Cliente WHERE Id = @IdCliente";
-                SqlCommand miComandoSql = new SqlCommand(consulta, miConexionSql);
-                miConexionSql.Open();
-                miComandoSql.Parameters.AddWithValue("IdCliente", ListaDeClientes.SelectedValue);
-                miComandoSql.ExecuteNonQuery();
-            }
-            catch (Exception ex)
-            {
-                MessageBox.Show(ex.ToString());
-            }
-            finally
-            {
-                miConexionSql.Close();
-                muestraDeLosClientes();
-                MessageBox.Show($"Has borrado ese cliente con exito");
+                try
+                {
+                    // creamos una consulta parametrica para hacer posible la eliminacion de un registro
+                    string consulta = "DELETE FROM Cliente WHERE Id = @IdCliente";
+                    SqlCommand miComandoSql = new SqlCommand(consulta, miConexionSql);
+                    miConexionSql.Open();
+                    miComandoSql.Parameters.AddWithValue("IdCliente", ListaDeClientes.SelectedValue);
+                    miComandoSql.ExecuteNonQuery();
+                }
+                catch (Exception ex)
+                {
+                    MessageBox.Show(ex.ToString());
+                }
+                finally
+                {
+                    miConexionSql.Close();
+                    muestraDeLosClientes();
+                    MessageBox.Show($"Has borrado ese cliente con exito");
+                }
             }
         }
 
@@ -231,6 +235,18 @@ namespace MiPrimerCRUD
                     MessageBox.Show(ex.ToString());
                 }
             }
+        }
+
+        private void BtnAyuda_Click(object sender, RoutedEventArgs e)
+        {
+            MessageBox.Show("Estas son las funciones de esta ventana:\n\n1. Agregar: Tienes que llenar todos los campos de texto que te aparecen, luego le das click al boton de agregar y con eso ya se registraria el cliente en la listbox" +
+                "\n\n2. Actualizar: Para actualizar un cliente tienes que seleccionarlo " +
+                "en la listBox y luego darle al boton de actualizar, luego se abrira una ventana con toda la informacion del cliente seleccionado para poderla editar, cuando ya llenes toda la nueva informacion " +
+                "del cliente, tienes que darle al boton de actualizar y te saldra una ventana preguntandote si quieres actualizar la informacion del cliente, si le das al boton de 'si' se cerrara la ventana y te dirigira " +
+                "nuevamente a la ventana de clientes donde ya tendras al cliente con la nueva informacion.\n\n3. Borrar: Para borrar un cliente tienes que seleccionarlo y darle al boton de Borrar, se te mostrara un mensaje " +
+                "preguntandote si de verdad quieres eliminarlo, y si le das que 'si' el cliente sera borrado.\n\n4. Generar Reporte: Para generar un reporte individual tienes que seleccionar a un cliente y darle al boton de generar " +
+                "reporte y automaticamente te generara un archivo de texto con la informacion del cliente seleccionado, ya si quieres reporte de todos los clientes, simplemente no selecciones a ninguno y dale al boton de generar reporte " +
+                "y con eso ya tendrias un nuevo archivo de texto con la informacion de todos los clientes que aparezcan en la listbox.");
         }
     }
 }
